+++ attributes/default.rb
@@ -31,6 +31,7 @@ default_unless['chef_client']['handler']['slack']['webhooks']['name'] = []
 # default['chef_client']['handler']['slack']['webhooks']['webhook1']['send_start_message'] = nil
 # default['chef_client']['handler']['slack']['webhooks']['webhook1']['message_detail_level'] = nil
 # default['chef_client']['handler']['slack']['webhooks']['webhook1']['cookbook_detail_level'] = nil
+# default['chef_client']['handler']['slack']['webhooks']['webhook1']['send_environment'] = nil
 
 # shared attributes
 default['chef_client']['handler']['slack']['timeout']    = 15
@@ -46,3 +47,5 @@ default['chef_client']['handler']['slack']['cookbook_detail_level'] = 'off'
 default['chef_client']['handler']['slack']['fail_only'] = false
 # Whether to send a message to when the Chef run starts
 default['chef_client']['handler']['slack']['send_start_message'] = false
+# Whether to send a message the node.chef_environment as well as the node.name
+default['chef_client']['handler']['slack']['send_environment'] = false